Gathered string methods into one file.
