+++ Makefile.docker
@@ -70,12 +70,12 @@ LIBFLAGS_OS4	= +aosppc -lamiga -lauto -o
 ##########################################################################
 # Object files which are part of iGame
 ##########################################################################
-OBJS		= src/funcs.o src/iGameGUI.o src/iGameMain.o src/strcasestr.o src/strdup.o src/iGame_cat.o
-OBJS_030	= src/funcs_030.o src/iGameGUI_030.o src/iGameMain_030.o src/strcasestr_030.o src/strdup_030.o src/iGame_cat_030.o
-OBJS_040	= src/funcs_040.o src/iGameGUI_040.o src/iGameMain_040.o src/strcasestr_040.o src/strdup_040.o src/iGame_cat_040.o
-OBJS_060	= src/funcs_060.o src/iGameGUI_060.o src/iGameMain_060.o src/strcasestr_060.o src/strdup_060.o src/iGame_cat_060.o
-OBJS_MOS	= src/funcs_MOS.o src/iGameGUI_MOS.o src/iGameMain_MOS.o src/strcasestr_MOS.o src/strdup_MOS.o src/iGame_cat_MOS.o
-OBJS_OS4	= src/funcs_OS4.o src/iGameGUI_OS4.o src/iGameMain_OS4.o src/strcasestr_OS4.o src/strdup_OS4.o src/iGame_cat_OS4.o
+OBJS		= src/funcs.o src/iGameGUI.o src/iGameMain.o src/strFuncs.o src/iGame_cat.o
+OBJS_030	= src/funcs_030.o src/iGameGUI_030.o src/iGameMain_030.o src/strFuncs_030.o src/iGame_cat_030.o
+OBJS_040	= src/funcs_040.o src/iGameGUI_040.o src/iGameMain_040.o src/strFuncs_040.o src/iGame_cat_040.o
+OBJS_060	= src/funcs_060.o src/iGameGUI_060.o src/iGameMain_060.o src/strFuncs_060.o src/iGame_cat_060.o
+OBJS_MOS	= src/funcs_MOS.o src/iGameGUI_MOS.o src/iGameMain_MOS.o src/strFuncs_MOS.o src/iGame_cat_MOS.o
+OBJS_OS4	= src/funcs_OS4.o src/iGameGUI_OS4.o src/iGameMain_OS4.o src/strFuncs_OS4.o src/iGame_cat_OS4.o
 
 ##########################################################################
 # Rule for building
@@ -127,11 +127,8 @@ src/iGameGUI.o: src/iGameGUI.c src/iGameGUI.h src/version.h src/iGame_cat.h
 src/iGameMain.o: src/iGameMain.c
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/iGameMain.c
 
-src/strcasestr.o: src/strcasestr.c
-	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/strcasestr.c
-
-src/strdup.o: src/strdup.c
-	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/strdup.c
+src/strFuncs.o: src/strFuncs.c
+	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/strFuncs.c
 
 src/iGame_cat.o: src/iGame_cat.c
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/iGame_cat.c
@@ -149,11 +146,8 @@ src/iGameGUI_030.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
 src/iGameMain_030.o: src/iGameMain.c
 	$(CC) $(CFLAGS_030) $(INCLUDES) -o $@ src/iGameMain.c
 
-src/strcasestr_030.o: src/strcasestr.c
-	$(CC) $(CFLAGS_030) $(INCLUDES) -o $@ src/strcasestr.c
-
-src/strdup_030.o: src/strdup.c
-	$(CC) $(CFLAGS_030) $(INCLUDES) -o $@ src/strdup.c
+src/strFuncs_030.o: src/strFuncs.c
+	$(CC) $(CFLAGS_030) $(INCLUDES) -o $@ src/strFuncs.c
 
 src/iGame_cat_030.o: src/iGame_cat.c
 	$(CC) $(CFLAGS_030) $(INCLUDES) -o $@ src/iGame_cat.c
@@ -171,11 +165,8 @@ src/iGameGUI_040.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
 src/iGameMain_040.o: src/iGameMain.c
 	$(CC) $(CFLAGS_040) $(INCLUDES) -o $@ src/iGameMain.c
 
-src/strcasestr_040.o: src/strcasestr.c
-	$(CC) $(CFLAGS_040) $(INCLUDES) -o $@ src/strcasestr.c
-
-src/strdup_040.o: src/strdup.c
-	$(CC) $(CFLAGS_040) $(INCLUDES) -o $@ src/strdup.c
+src/strFuncs_040.o: src/strFuncs.c
+	$(CC) $(CFLAGS_040) $(INCLUDES) -o $@ src/strFuncs.c
 
 src/iGame_cat_040.o: src/iGame_cat.c
 	$(CC) $(CFLAGS_040) $(INCLUDES) -o $@ src/iGame_cat.c
@@ -193,11 +184,8 @@ src/iGameGUI_060.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
 src/iGameMain_060.o: src/iGameMain.c
 	$(CC) $(CFLAGS_060) $(INCLUDES) -o $@ src/iGameMain.c
 
-src/strcasestr_060.o: src/strcasestr.c
-	$(CC) $(CFLAGS_060) $(INCLUDES) -o $@ src/strcasestr.c
-
-src/strdup_060.o: src/strdup.c
-	$(CC) $(CFLAGS_060) $(INCLUDES) -o $@ src/strdup.c
+src/strFuncs_060.o: src/strFuncs.c
+	$(CC) $(CFLAGS_060) $(INCLUDES) -o $@ src/strFuncs.c
 
 src/iGame_cat_060.o: src/iGame_cat.c
 	$(CC) $(CFLAGS_060) $(INCLUDES) -o $@ src/iGame_cat.c
@@ -215,11 +203,8 @@ src/iGameGUI_MOS.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
 src/iGameMain_MOS.o: src/iGameMain.c
 	$(CC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/iGameMain.c
 
-src/strcasestr_MOS.o: src/strcasestr.c
-	$(CC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/strcasestr.c
-
-src/strdup_MOS.o: src/strdup.c
-	$(CC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/strdup.c
+src/strFuncs_MOS.o: src/strFuncs.c
+	$(CC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/strFuncs.c
 
 src/iGame_cat_MOS.o: src/iGame_cat.c
 	$(CC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/iGame_cat.c
@@ -237,11 +222,8 @@ src/iGameGUI_OS4.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
 src/iGameMain_OS4.o: src/iGameMain.c
 	$(CC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/iGameMain.c
 
-src/strcasestr_OS4.o: src/strcasestr.c
-	$(CC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/strcasestr.c
-
-src/strdup_OS4.o: src/strdup.c
-	$(CC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/strdup.c
+src/strFuncs_OS4.o: src/strFuncs.c
+	$(CC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/strFuncs.c
 
 src/iGame_cat_OS4.o: src/iGame_cat.c
 	$(CC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/iGame_cat.c
@@ -251,7 +233,7 @@ src/iGame_cat_OS4.o: src/iGame_cat.c
 ##########################################################################
 
 clean:
-	rm iGame iGame.* src/funcs*.o src/iGameGUI*.o src/iGameMain*.o src/strcasestr*.o src/strdup*.o src/iGame_cat*.o $(catalog_files)
+	rm iGame iGame.* src/funcs*.o src/iGameGUI*.o src/iGameMain*.o src/strFuncs*.o src/iGame_cat*.o $(catalog_files)
 
 release: $(catalog_files)
 	cp required_files iGame-$(DRONE_TAG) -r
